created the architecture and corrected the name of the list
