+++ lib/features/zhurnal/widgets/point_list.dart
@@ -1,0 +1,26 @@
+final List<Map<String, dynamic>> pointList = [
+  {
+    "date": "11.12",
+    "name": "Математика",
+    "workname": "үй тапсырмасы",
+    "number": "5",
+  },
+  {
+    "date": "15.12",
+    "name": "Физика",
+    "workname": "жұптық жұмыс",
+    "number": "4",
+  },
+  {
+    "date": "13.12",
+    "name": "Химия",
+    "workname": "зертханалық жұмыс",
+    "number": "5",
+  },
+  {
+    "date": "18.12",
+    "name": "Математика",
+    "workname": "home work",
+    "number": "5",
+  },
+];
