+++ lib/features/zhurnal/widgets/profile_list.dart
@@ -1,0 +1,22 @@
+final List<Map<String, dynamic>> profileList = [
+  {
+    "date": "11.12",
+    "name": "Математика",
+    "profile": "сабақта жоқ",
+  },
+  {
+    "date": "15.12",
+    "name": "Физика",
+    "profile": "сабақта жоқ",
+  },
+  {
+    "date": "13.12",
+    "name": "Химия",
+    "profile": "сабақта жоқ",
+  },
+  {
+    "date": "18.12",
+    "name": "Математика",
+    "profile": "сабақта жоқ",
+  },
+];
